{"schema":"decodex/commit/1","summary":"Modularize CLI runtime and status boundaries","authority":"manual"}
