+++ apps/decodex/src/orchestrator/execution/resume.rs
@@ -1,0 +1,36 @@
+use crate::{
+	orchestrator::{CONTINUATION_PENDING_RUN_STATUS, IssueRunPlan, Result, RunSummary, StateStore},
+	state,
+};
+
+pub(crate) fn persist_issue_run_outcome(
+	state_store: &StateStore,
+	run_id: &str,
+	summary: &RunSummary,
+) -> Result<()> {
+	state_store.update_run_status(
+		run_id,
+		if summary.continuation_pending { CONTINUATION_PENDING_RUN_STATUS } else { "succeeded" },
+	)
+}
+
+pub(crate) fn resolve_resume_thread_id(
+	state_store: &StateStore,
+	issue_run: &IssueRunPlan,
+) -> Result<Option<String>> {
+	if let Some(run_attempt) = state_store.run_attempt(&issue_run.run_id)?
+		&& run_attempt.attempt_number() == issue_run.attempt_number
+		&& let Some(thread_id) = run_attempt.thread_id()
+	{
+		return Ok(Some(thread_id.to_owned()));
+	}
+
+	let marker = state::read_run_activity_marker_snapshot(&issue_run.worktree.path)?;
+
+	Ok(marker
+		.filter(|marker| {
+			marker.run_id() == issue_run.run_id
+				&& marker.attempt_number() == issue_run.attempt_number
+		})
+		.and_then(|marker| marker.thread_id().map(str::to_owned)))
+}
