+++ apps/decodex/src/orchestrator/execution/runtime_context.rs
@@ -1,0 +1,21 @@
+use crate::{
+	orchestrator::{GhPullRequestReviewStateInspector, Path, Result, ServiceConfig},
+	tracker,
+};
+
+pub(crate) fn configured_public_projection_privacy_classifier(
+	project: &ServiceConfig,
+) -> Result<tracker::privacy_classifier::ConfiguredPublicProjectionPrivacyClassifier> {
+	tracker::privacy_classifier::ConfiguredPublicProjectionPrivacyClassifier::from_config(
+		project.privacy_classifier(),
+	)
+}
+
+pub(crate) fn build_closeout_review_state_inspector(
+	project: &ServiceConfig,
+) -> GhPullRequestReviewStateInspector {
+	GhPullRequestReviewStateInspector {
+		github_token_env_var: Some(project.github().token_env_var().to_owned()),
+		github_command_path: project.github().command_path().map(Path::to_path_buf),
+	}
+}
